refactor: apply code review comments from prev. PR
