+++ priv/perf/apps/load_test/lib/service/nonce_tracker.ex
@@ -20,16 +20,16 @@ defmodule LoadTest.Service.NonceTracker do
 
   alias ExPlasma.Encoding
 
-  def start_link(init_state \\ %{}) do
-    GenServer.start_link(__MODULE__, init_state, name: __MODULE__)
-  end
-
   def update_nonce(addr) do
     GenServer.call(__MODULE__, {:update_nonce, addr}, :infinity)
   end
 
-  def init(init_state) do
-    {:ok, init_state}
+  def start_link(args) do
+    GenServer.start_link(__MODULE__, args, name: __MODULE__)
+  end
+
+  def init(_) do
+    {:ok, Map.new()}
   end
 
   def handle_call({:update_nonce, addr}, _from, state) do
