+++ priv/perf/apps/load_test/mix.exs
@@ -19,7 +19,8 @@ defmodule LoadTest.MixProject do
   # Run "mix help compile.app" to learn about applications.
   def application do
     [
-      extra_applications: [:logger]
+      extra_applications: [:logger],
+      mod: {LoadTest.Application, []}
     ]
   end
